Update VertexHeightMap16 to work in KSP 1.9.1.
While VertexHeight PQS Mods themselves appear to not require any changes to update to KSP 1.9.1 (for instance, the VertexHeightDeformity expansion can be re-compiled for KSP 1.9.1 without any changes required to its code), the way KSP loads maps was changed in 1.8. Previously, the VertexHeightMap16 mod would use the alpha color channel to store the lower 8-bit of the 16-bit height value, but KSP now discards the value in the alpha channel when loading heightmaps. Therefore, I changed the code to instead use the blue and green color channels to store the height information. I also updated the converter program to use this new format.