Fix VHM16 breaking when loaded while a Kopernicus cache file for the celestial body is present.
Somehow, Kopernicus likes to try to sample the heightmap even though it hasn't been loaded yet, but for some reason also only if a cache file for the celestial body in question exists and has been loaded (when 'update' in the 'Debug' node of the Kopernicus config is false). By simply return 0 as the sample if this is the case, things return to normal.
The polar spike/polar hole bug from the base game also turned out to be fixable for VHM16 and VHM24, so this is done in this commit as well.